cleanup code for tag management
